Clean up the ranking code, make it serial.
The serial ranking code use the same qlg_matches table when it decides
to rank. We have thus opened up the path where we can runs a series of
rankings in serial and then finally decide what we want to do with the
resulting rankings. We probably want to output it in a CSV file, but
that is for later.